+++ config/config.exs
@@ -1,7 +1,7 @@
 import Config
 
 config :dart_sass,
-  version: "1.39.0",
+  version: "1.43.4",
   another: [
     args: ["--version"]
   ]