+++ config/config.exs
@@ -1,7 +1,7 @@
 import Config
 
 config :dart_sass,
-  version: "1.43.4",
+  version: "1.49.0",
   another: [
     args: ["--version"]
   ]