+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 ## v0.3.1
 
+  * Update Sass version to `1.49.0`
   * Attach system target architecture to saved esbuild executable (h/t @cw789)
   * Use user cache directory (h/t @josevalim)
   * Add support for 32bit linux (h/t @derek-zhou)
   * Support `HTTP_PROXY/HTTPS_PROXY` to fetch esbuild (h/t @iaddict)
-  * Update Sass version to `1.43.4`
   * Fallback to _build if Mix.Project is not available
   * Allow `config :dart_sass, :path, path` to configure the path to the Sass executable (or snapshot)
   * Support OTP 24 on Apple M1 architectures (via Rosetta2)